+++ core/src/main/java/com/srotya/sidewinder/core/api/InfluxApi.java
@@ -15,7 +15,6 @@
  */
 package com.srotya.sidewinder.core.api;
 
-import java.io.IOException;
 import java.util.List;
 
 import javax.ws.rs.BadRequestException;
@@ -29,7 +28,7 @@ import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
 import com.srotya.sidewinder.core.monitoring.MetricsRegistryService;
 import com.srotya.sidewinder.core.rpc.Point;
-import com.srotya.sidewinder.core.storage.StorageEngine;
+import com.srotya.sidewinder.core.storage.processor.PointProcessor;
 import com.srotya.sidewinder.core.utils.InfluxDecoder;
 
 /**
@@ -38,11 +37,11 @@ import com.srotya.sidewinder.core.utils.InfluxDecoder;
 @Path("/influx")
 public class InfluxApi {
 
-	private StorageEngine storageEngine;
 	private Meter meter;
+	private PointProcessor proc;
 
-	public InfluxApi(StorageEngine storageEngine) {
-		this.storageEngine = storageEngine;
+	public InfluxApi(PointProcessor proc) {
+		this.proc = proc;
 		MetricRegistry registry = MetricsRegistryService.getInstance().getInstance("requests");
 		meter = registry.meter("influx-writes");
 	}
@@ -61,8 +60,8 @@ public class InfluxApi {
 		meter.mark(dps.size());
 		for (Point dp : dps) {
 			try {
-				storageEngine.writeDataPoint(dp, preSorted);
-			} catch (IOException e) {
+				proc.writeDataPoint(dbName, dp);
+			} catch (Exception e) {
 				e.printStackTrace();
 				throw new BadRequestException(e);
 			}
