+++ core/src/main/java/com/srotya/sidewinder/core/storage/processor/PointProcessor.java
@@ -1,0 +1,29 @@
+/**
+ * Copyright 2018 Ambud Sharma
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ * 		http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.srotya.sidewinder.core.storage.processor;
+
+import java.io.IOException;
+
+import com.srotya.sidewinder.core.rpc.Point;
+
+/**
+ * @author ambud
+ */
+public interface PointProcessor {
+	
+	public void writeDataPoint(String dbName, Point point) throws InterruptedException, IOException;
+
+}
