+++ ingesters/src/main/java/com/srotya/sidewinder/ingesters/influx/HTTPDataPointDecoder.java
@@ -136,7 +136,7 @@ public class HTTPDataPointDecoder extends SimpleChannelInboundHandler<Object> {
 						meter.inc(dps.size());
 						for (Point dp : dps) {
 							try {
-								engine.writeDataPoint(dp, false);
+								engine.writeDataPointLocked(dp, false);
 								logger.fine("Accepted:" + dp + "\t" + new Date(dp.getTimestamp()));
 							} catch (IOException e) {
 								logger.fine("Dropped:" + dp + "\t" + e.getMessage());
